refactor(example): use barrel import for book module (#1850)
Closes #1684
