+++ projects/example-app/src/app/books/actions/selected-book-page.actions.ts
@@ -1,5 +1,6 @@
 import { createAction, props } from '@ngrx/store';
-import { Book } from '@example-app/books/models/book';
+
+import { Book } from '@example-app/books/models';
 
 /**
  * Add Book to Collection Action
