+++ projects/example-app/src/app/books/containers/collection-page.component.ts
@@ -1,9 +1,10 @@
 import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
+
 import { select, Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
 
 import { CollectionPageActions } from '@example-app/books/actions';
-import { Book } from '@example-app/books/models/book';
+import { Book } from '@example-app/books/models';
 import * as fromBooks from '@example-app/books/reducers';
 
 @Component({
