+++ projects/example-app/src/app/books/containers/selected-book-page.component.ts
@@ -1,10 +1,11 @@
-import { Component, ChangeDetectionStrategy } from '@angular/core';
-import { Store, select } from '@ngrx/store';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+import { select, Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
 
-import * as fromBooks from '@example-app/books/reducers';
 import { SelectedBookPageActions } from '@example-app/books/actions';
-import { Book } from '@example-app/books/models/book';
+import { Book } from '@example-app/books/models';
+import * as fromBooks from '@example-app/books/reducers';
 
 @Component({
   selector: 'bc-selected-book-page',
