+++ projects/example-app/src/app/books/containers/view-book-page.component.spec.ts
@@ -1,17 +1,20 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { Store } from '@ngrx/store';
 import { ActivatedRoute } from '@angular/router';
+
+import { Store } from '@ngrx/store';
+import { provideMockStore, MockStore } from '@ngrx/store/testing';
 import { BehaviorSubject } from 'rxjs';
 import { MatCardModule } from '@angular/material';
 
-import { ViewBookPageComponent } from '@example-app/books/containers/view-book-page.component';
+import {
+  BookAuthorsComponent,
+  BookDetailComponent,
+} from '@example-app/books/components';
+import { SelectedBookPageComponent } from '@example-app/books/containers';
+import { ViewBookPageComponent } from '@example-app/books/containers';
 import { ViewBookPageActions } from '@example-app/books/actions';
 import * as fromBooks from '@example-app/books/reducers';
-import { SelectedBookPageComponent } from '@example-app/books/containers/selected-book-page.component';
-import { BookDetailComponent } from '@example-app/books/components/book-detail.component';
-import { BookAuthorsComponent } from '@example-app/books/components/book-authors.component';
 import { AddCommasPipe } from '@example-app/shared/pipes/add-commas.pipe';
-import { provideMockStore, MockStore } from '@ngrx/store/testing';
 
 describe('View Book Page', () => {
   let fixture: ComponentFixture<ViewBookPageComponent>;
