+++ projects/example-app/src/app/books/effects/book.effects.spec.ts
@@ -1,16 +1,17 @@
 import { TestBed } from '@angular/core/testing';
+
 import { Actions } from '@ngrx/effects';
 import { provideMockActions } from '@ngrx/effects/testing';
 import { cold, getTestScheduler, hot } from 'jasmine-marbles';
 import { Observable } from 'rxjs';
 
-import { GoogleBooksService } from '@example-app/core/services/google-books.service';
 import {
   BooksApiActions,
   FindBookPageActions,
 } from '@example-app/books/actions';
-import { Book } from '@example-app/books/models/book';
-import { BookEffects } from '@example-app/books/effects/book.effects';
+import { BookEffects } from '@example-app/books/effects';
+import { Book } from '@example-app/books/models';
+import { GoogleBooksService } from '@example-app/core/services/google-books.service';
 
 describe('BookEffects', () => {
   let effects: BookEffects;
