+++ projects/example-app/src/app/books/effects/collection.effects.spec.ts
@@ -1,20 +1,20 @@
 import { TestBed } from '@angular/core/testing';
+
+import {
+  CollectionApiActions,
+  CollectionPageActions,
+  SelectedBookPageActions,
+} from '@example-app/books/actions';
+import { CollectionEffects } from '@example-app/books/effects';
+import { Book } from '@example-app/books/models';
+import {
+  BookStorageService,
+  LOCAL_STORAGE_TOKEN,
+} from '@example-app/core/services';
 import { Actions } from '@ngrx/effects';
 import { provideMockActions } from '@ngrx/effects/testing';
 import { cold, hot } from 'jasmine-marbles';
 import { Observable } from 'rxjs';
-
-import {
-  CollectionApiActions,
-  SelectedBookPageActions,
-  CollectionPageActions,
-} from '@example-app/books/actions';
-import { Book } from '@example-app/books/models/book';
-import { CollectionEffects } from '@example-app/books/effects/collection.effects';
-import {
-  BookStorageService,
-  LOCAL_STORAGE_TOKEN,
-} from '@example-app/core/services/';
 
 describe('CollectionEffects', () => {
   let db: any;
