+++ projects/example-app/src/app/books/effects/index.ts
@@ -1,0 +1,2 @@
+export * from './book.effects';
+export * from './collection.effects';
