+++ projects/example-app/src/app/books/guards/index.ts
@@ -1,0 +1,1 @@
+export * from './book-exists.guard';
