+++ projects/example-app/src/app/books/index.ts
@@ -1,0 +1,1 @@
+export * from './books.module';
