+++ projects/example-app/src/app/books/models/index.ts
@@ -1,0 +1,1 @@
+export * from './book';
