+++ projects/example-app/src/app/core/services/book-storage.service.ts
@@ -1,12 +1,9 @@
-import {
-  Injectable,
-  InjectionToken,
-  FactoryProvider,
-  Inject,
-} from '@angular/core';
+import { Inject, Injectable, InjectionToken } from '@angular/core';
+
 import { Observable, of, throwError } from 'rxjs';
 import { map, tap } from 'rxjs/operators';
-import { Book } from '@example-app/books/models/book';
+
+import { Book } from '@example-app/books/models';
 
 export function storageFactory() {
   return typeof window === undefined || typeof localStorage === undefined
