+++ projects/example-app/src/app/core/services/google-books.service.ts
@@ -1,9 +1,10 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
+
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 
-import { Book } from '@example-app/books/models/book';
+import { Book } from '@example-app/books/models';
 
 @Injectable({
   providedIn: 'root',
